Deplacement du main() dans rush-hour.c | Création du module interface_txt | Quelques modifications mineures
